+++ src/adapters/bun.ts
@@ -5,9 +5,9 @@ import { createMiddleware } from "hono/factory";
 import { logger } from "hono/logger";
 import type { BlankEnv } from "hono/types";
 import { type ServerBuild, createRequestHandler } from "react-router";
+import { cleanUpgradeListeners, createWebSocket, patchUpgradeListener } from "../helpers";
 import { cache } from "../middleware";
 import type { HonoServerOptionsBase, WithWebsocket, WithoutWebsocket } from "../types/hono-server-options-base";
-import { cleanUpgradeListeners, createWebSocket, patchUpgradeListener } from "../utils";
 
 type CustomBunServer = Serve &
   ServeOptions & {
